refactor cmd. flags overwrite config if set
